% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMD.R
\name{getAMatrix}
\alias{getAMatrix}
\title{Create a Koopman Matrix}
\usage{
getAMatrix(data, p = 1)
}
\arguments{
\item{data}{the matrix data for which you wish to predict future columns}

\item{p}{The Percentage of explanation of the SVD required (default=1)
Returns an object of class kdmd}
}
\description{
Create a Koopman Matrix
}
\examples{
## Not run:
m<-matrix(ncol=10,nrow=3)
m[1,]<-1:10
m[2,]<-2:11
m[3,]<-3,12
A<-getAmatrix(m)
}
