% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMD.R
\name{predict}
\alias{predict}
\title{Predict from a Koopman Matrix}
\usage{
predict(A, data, l, p = 1)
}
\arguments{
\item{A}{the Koopman Matrix for prediction}

\item{data}{the matrix data for which you wish to predict future columns}

\item{l}{the length of the predictions (number of columns to predict)}

\item{p}{The Percentage of explanation of the SVD required (default=1)
Returns a matrix with l additional columns}
}
\description{
Predict from a Koopman Matrix
}
\examples{
## Not run:
m<-matrix(ncol=10,nrow=3)
m<-matrix(ncol=10,nrow=3)
m[1,]<-1:10
m[2,]<-2:11
m[3,]<-3:12
A<-getAmatrix(m)
predict(A,m,1)
}
