% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMD.R
\name{predict.kdmd}
\alias{predict.kdmd}
\title{Predict from a Koopman Matrix}
\usage{
\method{predict}{kdmd}(A, data, l = 1)
}
\arguments{
\item{A}{the Koopman Matrix for prediction}

\item{data}{the matrix data for which you wish to predict future columns - must be conformable to A}

\item{l}{the length of the predictions (number of columns to predict - default=1)
Returns a matrix with l additional columns}
}
\description{
Predict from a Koopman Matrix
}
\examples{
## Not run:
m<-matrix(ncol=10,nrow=3)
m<-matrix(ncol=10,nrow=3)
m[1,]<-1:10
m[2,]<-2:11
m[3,]<-3:12
A<-getAMatrix(m)
predict(A,m,1)
}
